+++ src/html/generator.ml
@@ -333,7 +333,7 @@ and items ~config ~resolve l : item Html.elt list =
           | `Inline -> doc @ included_html
           | `Closed -> details ~open':false
           | `Open -> details ~open':true
-          | `Default -> details ~open':(Config.open_details config)
+          | `Default -> details ~open':config.open_details
         in
         (continue_with [@tailcall]) rest content
     | Declaration { Item.attr; anchor; content; doc } :: rest ->
@@ -375,6 +375,30 @@ module Toc = struct
     List.map section toc
 end
 
+module Breadcrumbs = struct
+  open Types
+
+  let gen_breadcrumbs ~config ~url =
+    let rec get_parent_paths x =
+      match x with
+      | [] -> []
+      | x :: xs -> (
+          match Odoc_document.Url.Path.of_list (List.rev (x :: xs)) with
+          | Some x -> x :: get_parent_paths xs
+          | None -> get_parent_paths xs)
+    in
+    let to_breadcrumb breadcrumb_url =
+      let href =
+        Link.href ~config ~resolve:(Current url)
+          (Odoc_document.Url.from_path breadcrumb_url)
+      in
+      let name = breadcrumb_url.name in
+      { href; name }
+    in
+    get_parent_paths (List.rev (Odoc_document.Url.Path.to_list url))
+    |> List.rev |> List.map to_breadcrumb
+end
+
 module Page = struct
   let on_sub = function
     | `Page _ -> None
@@ -385,7 +409,7 @@ module Page = struct
 
   let rec include_ ~config { Subpage.content; _ } = page ~config content
 
-  and subpages ~config subpages =
+  and subpages ~(config : Config.Html_page.t) subpages =
     Utils.list_concat_map ~f:(include_ ~config) subpages
 
   and page ~config p : Odoc_document.Renderer.page list =
@@ -398,15 +422,52 @@ module Page = struct
     let resolve = Link.Current url in
     let i = Doctree.Shift.compute ~on_sub i in
     let uses_katex = Doctree.Math.has_math_elements p in
-    let toc = Toc.gen_toc ~config ~resolve ~path:url i in
+    let toc = Toc.gen_toc ~config:config.base ~resolve ~path:url i in
+    let breadcrumbs = Breadcrumbs.gen_breadcrumbs ~config:config.base ~url in
+    let content = (items ~config:config.base ~resolve i :> any Html.elt list) in
     let header =
-      items ~config ~resolve (Doctree.PageTitle.render_title p @ preamble)
+      items ~config:config.base ~resolve
+        (Doctree.PageTitle.render_title p @ preamble)
     in
-    let content = (items ~config ~resolve i :> any Html.elt list) in
-    Tree.make ~config ~header ~toc ~url ~uses_katex url.name content subpages
+    Html_page.make ~config ~header ~toc ~breadcrumbs ~url ~uses_katex title
+      content subpages
 end
 
 let render ~config page = Page.page ~config page
+
+module Meta = struct
+  let on_sub = function
+    | `Page _ -> None
+    | `Include x -> (
+        match x.Include.status with
+        | `Closed | `Open | `Default -> None
+        | `Inline -> Some 0)
+
+  let rec include_ ~config { Subpage.content; _ } = page ~config content
+
+  and subpages ~(config : Config.Base.t) subpages =
+    Utils.list_concat_map ~f:(include_ ~config) subpages
+
+  and page ~config p : Odoc_document.Renderer.page list =
+    let { Odoc_document.Types.Page.title; kind; preamble; items = i; url } =
+      Doctree.Labels.disambiguate_page p
+    and subpages =
+      (* Don't use the output of [disambiguate_page] to avoid unecessarily
+         mangled labels. *)
+      subpages ~config @@ Doctree.Subpages.compute p
+    in
+    let resolve = Link.Current url in
+    let i = Doctree.Shift.compute ~on_sub i in
+    let uses_katex = Doctree.Math.has_math_elements p in
+    let toc = Toc.gen_toc ~config ~resolve ~path:url i in
+    let breadcrumbs = Breadcrumbs.gen_breadcrumbs ~config ~url in
+    let content = (items ~config ~resolve i :> any Html.elt list) in
+    Html_fragment_json.make ~config
+      ~preamble:(items ~config ~resolve preamble :> any Html.elt list)
+      ~breadcrumbs ~toc ~url ~kind ~uses_katex ~title content subpages
+end
+
+let render_fragment_with_meta ~config page = Meta.page ~config page
 
 let doc ~config ~xref_base_uri b =
   let resolve = Link.Base xref_base_uri in