+++ src/html/html_page.mli
@@ -21,9 +21,10 @@ module Html = Tyxml.Html
 (** {1 Page creator} *)
 
 val make :
-  config:Config.t ->
+  config:Config.Html_page.t ->
   url:Odoc_document.Url.Path.t ->
   header:Html_types.flow5_without_header_footer Html.elt list ->
+  breadcrumbs:Types.breadcrumb list ->
   toc:Types.toc list ->
   uses_katex:bool ->
   string ->