+++ src/html/types.ml
@@ -8,3 +8,5 @@ type toc = {
   href : string;
   children : toc list;
 }
+
+type breadcrumb = { href : string; name : string }